refactor: CycleRepository에서 동적 쿼리 분리
