+++ backend/smody/src/test/java/com/woowacourse/smody/cycle/repository/CycleRepositoryTest.java
@@ -25,6 +25,8 @@ class CycleRepositoryTest extends RepositoryTest {
 
     @Autowired
     private CycleRepository cycleRepository;
+    @Autowired
+    private DynamicCycleRepository dynamicCycleRepository;
 
     @Autowired
     private ResourceFixture fixture;
@@ -84,7 +86,7 @@ class CycleRepositoryTest extends RepositoryTest {
         fixture.사이클_생성_SUCCESS(더즈_ID, 오늘의_운동_ID, now.plusSeconds(1L));
 
         // when
-        List<ChallengingRecord> actual = cycleRepository.findAllChallengingRecordByMemberAfterTime(조조그린_ID,
+        List<ChallengingRecord> actual = dynamicCycleRepository.findAllChallengingRecordByMemberAfterTime(조조그린_ID,
                 now.minusDays(3));
 
         for (ChallengingRecord challengingRecord : actual) {
