+++ backend/smody/src/test/java/com/woowacourse/smody/support/RepositoryTest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.smody.support;
 
+import com.woowacourse.smody.cycle.repository.DynamicCycleRepository;
+import com.woowacourse.smody.cycle.repository.DynamicCycleRepositoryImpl;
 import com.woowacourse.smody.support.config.QueryFactoryTestConfig;
 import com.woowacourse.smody.support.isoloation.SmodyDatabaseManager;
 import org.junit.jupiter.api.AfterEach;
@@ -9,7 +11,12 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
 @DataJpaTest(showSql = false)
-@Import({ResourceFixture.class, SmodyDatabaseManager.class, QueryFactoryTestConfig.class})
+@Import({
+        ResourceFixture.class,
+        SmodyDatabaseManager.class,
+        QueryFactoryTestConfig.class,
+        DynamicCycleRepositoryImpl.class
+})
 public class RepositoryTest {
 
     @Autowired
